refactor: separate Watch from StackedList
This will allow displaying a generic list of items, not only watchable
lists.
