+++ frontend/src/views/cluster/Nodes.vue
@@ -8,7 +8,7 @@ file, You can obtain one at http://mozilla.org/MPL/2.0/.
     <div class="px-3 py-2 mb-2">
       <t-breadcrumbs :active="$route.params.cluster + ' Nodes'"/>
     </div>
-    <stacked-list 
+    <watch
         class="flex-1"
         :resource="{type: 'nodes.v1'}"
         showCount
@@ -58,18 +58,18 @@ file, You can obtain one at http://mozilla.org/MPL/2.0/.
           </div>
         </router-link>
       </template>
-    </stacked-list>
+    </watch>
   </div>
 </template>
 
 <script type="ts">
 import { Options, Vue } from 'vue-class-component';
-import StackedList from '../../components/StackedList.vue';
+import Watch from '../../components/Watch.vue';
 import TBreadcrumbs from '../../components/TBreadcrumbs.vue';
 
 @Options({
   components: {
-    StackedList,
+    Watch,
     TBreadcrumbs,
   },
